Minor cleanup, s/activate/add
Consistent verbiage good.

I keep dancing around the work started in 28a9d33,
but the subtle differences between backing up a file when adding it
and backing up a file when updating it are... subtle, and I'd forgotten
about them. I feel like there's something cleaner in there, but:

- `add` has to hash the game file as it makes the backup,
  whereas `update` already hashed the game file.
  (That's why we're making a new backup!)

- This makes `add` use `hash_and_write(reader, temp_file)`
  while `update` can just use `fs::copy()`.

- `add` should freak out if there's already a backup file in
  `backup_path`, but this is what we expect in `update`.

Maybe I could at least centralize the last steps of moving completed
temps to the backup directory and clearing the (hopefully dir-only)
temp tree when we're done.