WIP: Add support back for Zip files via Piz.
That's the whole reason I wrote Piz. But using it here was quite a
struggle: I wanted to encapsulate everything we need (a memory map of
the archive, the `piz::ZipArchive`, the file tree from the archive,
and a reference to the base directory) in `ZipMod`, and turns out
self-referential structs are huge pain with Rust's lifetime system
without a bunch of `unsafe` partial initialization, pinning, or a crate
that does a bunch of that for you. I opted for #3 and went with
`owning_ref`, along with a (hopefully sound, knock on wood) cast from a
pointer to a `static`-lifetime reference.