Piz WIP
Current issue: I want to pack up a memory map, the ZipArchive that
depends on it, and the FileTree that depends on that ZipArchive, all in
one struct. The resulting struct would be self-referential, so now I'm
down a `Pin` rabbit hole.

Discussions like
https://users.rust-lang.org/t/how-do-i-create-self-referential-structures-with-pin/24982
and the Pin docs suggest using pointers like `NotNull<T>`, but the
trouble is that the references live _in_ ZipArchive, FileTree, etc.
Without writing those, I have to figure out what to do with their
lifetime params, i.e.,

    pub struct ZipMod {
        mmap: Mmap,
        z: ZipArchive<'what_do>, // references mmap
    }